requirements.txt
streamlit
python-dotenv
openai
langmem
langgraph
chromadb
sentence-transformers
langchain-chroma 
langchain-huggingface



streamlit
mem0ai
python-dotenv
tiktoken